ml/pyproject.toml
[tool.poetry]
name = "tuo_ml"
version = "0.0.0"
description = "Machine Learning for TUO"
authors = ["Alexander Puck Neuwirth <alexander@neuwirth-informatik.de>"]
readme = "README.md"
repository = "https://github.com/APN-Pucky/tyrant_optimize"

[tool.poetry.dependencies]
python = "^3.8"
deprecation = "*"

pyyaml="*"
tqdm="*"
numpy="^1.24"
scikit-learn="^1.1"
xgboost="^1.6"
sklearn2pmml="^0.92"
pandas="^1.5"
smpl_io="^1.1.0"

[tool.poetry.scripts]
tuo_train = 'tuo_ml.train:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
dirty = true
# modified dunami default pattern without v
#pattern = '(?x)(?# ignore whitespace)^((?P<epoch>\d+)!)?(?P<base>\d+(\.\d+)*)(?# v1.2.3 or v1!2000.1.2)([-._]?((?P<stage>[a-zA-Z]+)[-._]?(?P<revision>\d+)?))?(?# b0)(\+(?P<tagged_metadata>.+))?$(?# +linux)'
strict = true
format-jinja = "{% if distance == 0 %}{{ base }}{% else %}{{ base }}.{{ distance }}{% endif %}"
